feat: move Game model, GameRepository and HintLog model to :domain:core module
