+++ app/src/main/java/com/example/sudokuslayer/Module.kt
@@ -3,6 +3,7 @@ package com.example.sudokuslayer
 import com.example.data.core_android.coreAndroidModule
 import com.example.data.game.dataGameModule
 import com.example.data.settings.dataSettingsModule
+import com.example.domain.creator.domainCreatorModule
 import com.example.domain.game.domainGameModule
 import com.example.sudokuslayer.presentation.screen.game.gameModule
 import com.example.sudokuslayer.presentation.screen.sudokucreator.sudokuCreatorModule
@@ -17,5 +18,6 @@ val appModule =
 			dataGameModule,
 			sudokuCreatorModule,
 			domainGameModule,
+			domainCreatorModule,
 		)
 	}
