+++ data/game/src/main/java/com/example/data/game/AndroidProtoGameRepository.kt
@@ -7,10 +7,10 @@ import com.example.data.game.mappers.toProtoDifficulty
 import com.example.data.game.mappers.toProtoGame
 import com.example.data.game.mappers.toProtoGrid
 import com.example.data.game.mappers.toProtoHintLog
+import com.example.domain.core.Game
 import com.example.domain.core.GameDifficulty
-import com.example.domain.game.models.Game
-import com.example.domain.game.models.HintLog
-import com.example.domain.game.repositories.GameRepository
+import com.example.domain.core.GameRepository
+import com.example.domain.core.HintLog
 import com.example.sudoku.model.SudokuCellData
 import com.example.sudoku.model.SudokuGrid
 import kotlinx.coroutines.flow.Flow
