+++ domain/creator/src/main/java/com/example/domain/creator/Module.kt
@@ -1,0 +1,8 @@
+package com.example.domain.creator
+
+import org.koin.dsl.module
+
+val domainCreatorModule =
+	module {
+		factory { CreateNewGameUseCase() }
+	}
