+++ domain/game/src/main/java/com/example/domain/game/usecases/ResetGameUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.domain.game.usecases
 
-import com.example.domain.game.models.Game
+import com.example.domain.core.Game
 import com.example.domain.game.repositories.OperationRepository
 import com.example.sudoku.model.clearGrid
 import com.example.sudoku.model.clearMatchingNumberHighlight
